requirements.txt
torch
transformers
peft
accelerate
bitsandbytes
datasets
streamlit
Pillow
fastapi
uvicorn
diffusers
langchain
ollama
mlx
mlx-image